feat(EN-7797): profileCard option noHelp + CardList use grid css + design Directory page
