+++ src/components/backoffice/dashboard/DashboardRecommendationsCard/DashboardRecommendationsCard.styles.ts
@@ -1,7 +1,5 @@
 import styled from 'styled-components';
 
 export const StyledDashboardRecommendationsList = styled.div`
-  ul {
-    gap: 26px;
-  }
+  width: 100%;
 `;
