+++ src/components/backoffice/directory/Directory/Directory.styles.ts
@@ -1,8 +1,15 @@
 import styled from 'styled-components';
 
+export const StyledHeaderDirectory = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 35px;
+`;
+
 export const StyledDirectoryContainer = styled.div`
   display: flex;
   flex-direction: column;
+  gap: 50px;
 `;
 
 export const StyledDirectoryButtonContainer = styled.div`
