+++ src/components/backoffice/directory/Directory/index.ts
@@ -1,0 +1,1 @@
+export * from './Directory';
