--- src/components/backoffice/directory/DirectoryContainer/index.ts
@@ -1,1 +1,0 @@
-export * from './DirectoryContainer';
