+++ src/components/backoffice/referer/dashboard/DashboardReferedCandidateList/DashboardReferedCandidateList.styles.ts
@@ -1,7 +1,5 @@
 import styled from 'styled-components';
 
 export const StyledDashboardReferedCandidateList = styled.div`
-  ul {
-    gap: 26px;
-  }
+  width: 100%;
 `;
