+++ src/components/headers/HeaderBackoffice/HeaderBackoffice.styles.ts
@@ -25,3 +25,7 @@ export const StyledHeaderBackOffice = styled.div`
     margin: 30px 0;
   }
 `;
+
+export const StyledBackgroundedHeaderBackoffice = styled.div`
+  background-color: ${COLORS.lightGray};
+`;
