+++ src/components/utils/CardList/CardList.styles.ts
@@ -4,14 +4,14 @@ export const StyledCardListContainer = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 100%;
 `;
 
-export const StyledCardList = styled.ul`
-  list-style-type: none;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 56px;
-  justify-content: center;
+export const StyledCardList = styled.div`
+  display: grid;
+  width: 100%;
+  grid-template-columns: repeat(auto-fill, minmax(272px, 1fr));
+  gap: ${({ condensed }) => (condensed ? '20px' : '55px')};
   padding: 0;
 `;
