+++ src/components/utils/CardList/CardListItem/CardListItem.styles.ts
@@ -1,8 +1,7 @@
 import styled from 'styled-components';
 
-export const StyledCardListItem = styled.li`
+export const StyledCardListItem = styled.div`
   display: flex;
-
   > * {
     flex: 1;
     display: flex;
