+++ src/constants/styles.ts
@@ -11,7 +11,7 @@ export const HEIGHTS = {
   OFFER_ADMIN_INFO_HEIGHT: 176,
   OFFER_INFO_HEIGHT: 156,
   OFFER_CTA_HEIGHT: 55,
-  DEFAULT_SECTION_PADDING: 40,
+  DEFAULT_SECTION_PADDING: 50,
   SECTION_PADDING: 24,
   SECTION_PADDING_MOBILE: 12,
   MESSAGING_DESKTOP_BORDER_SIZE: 3,
