+++ src/pages/backoffice/annuaire.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import { LayoutBackOffice } from 'src/components/backoffice/LayoutBackOffice';
 import { LoadingScreen } from 'src/components/backoffice/LoadingScreen';
-import { DirectoryContainer } from 'src/components/backoffice/directory/DirectoryContainer/DirectoryContainer';
+import { Directory } from 'src/components/backoffice/directory/Directory/Directory';
 import { useDirectoryRoleRedirection } from 'src/components/backoffice/directory/useDirectoryRoleRedirection';
-import { HeaderBackoffice } from 'src/components/headers/HeaderBackoffice';
-import { Section } from 'src/components/utils';
 import { useRole } from 'src/hooks/queryParams/useRole';
 
 const Annuaire = () => {
@@ -14,15 +12,7 @@ const Annuaire = () => {
 
   return (
     <LayoutBackOffice title="Annuaire">
-      <Section className="custom-page">
-        <HeaderBackoffice
-          title="Bienvenue sur votre réseau"
-          description={
-            "Découvrez les membres de la communauté et développez votre carnet d'adresse."
-          }
-        />
-        {role ? <DirectoryContainer /> : <LoadingScreen />}
-      </Section>
+      {role ? <Directory /> : <LoadingScreen />}
     </LayoutBackOffice>
   );
 };
